THE INITIAL COMMIT
This project is a template for creating a React component for publishing
to the NPM register.

This has a "DemoComponent" that should be replaced with your actual
component. Bable and Webpack are pre-configured and support exists
for using Jest and Enzyme. There is a sample test that tests that the
Demo component loads.